+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Bug fixes
 
+-   Update the Python `Provider` class to use parameter naming consistent with other resources. (https://github.com/pulumi/pulumi-kubernetes/pull/1039).
 -   Change URN for apiregistration resources. (https://github.com/pulumi/pulumi-kubernetes/pull/1021).
 
 ## 1.5.8 (March 16, 2020)